Uploaded addon with custom changes
Changes made include:

Extending max players to select from from 15 to 20. This was done by adding MF16-20 in the same method that 1-15 were done.
Shaman class color so that there is distinction from paladins and shamans in the list. This is just by an extra elif statement where colors are assigned.

Known bugs: When making a new list, the 15th "healer" is under the new/save button. This exists without my extra changes.